Different approach to the arrays (this one works)
My earlier attempts were to rewrite the various loops to deal with the new arrays. That kepy causing problems, because there were many places where code needed to be changed, and it needed to be changed in different ways. Instead, I just load the JSON file and immediately separate it out into configs for each individual server, so the rest of the code is untouched.