Fix SGE parallel environment settings
The desired behaviour is for all nodes of parallel jobs to be on the same machine.
This requires the strategy $pe_slots, not $fill_up as it was set before. The total number of nodes has also been set to 9999 so that it is limited by the actual nodes available on each machine.